backend/requirements.txt
fastapi
uvicorn
pyyaml
httpx
beautifulsoup4
pydantic
python-multipart
